requirements.txt
absl-py==2.1.0+computecanada
aiohttp==3.8.4+computecanada
aiosignal==1.3.1+computecanada
albumentations==0.4.3
antlr4-python3-runtime==4.8+computecanada
async-timeout==4.0.3+computecanada
attrs==23.2.0+computecanada
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
einops==0.5.0
frozenlist==1.4.1
fsspec==2024.5.0+computecanada
ftfy==6.2.0
future==1.0.0+computecanada
google-auth==2.29.0
google-auth-oauthlib==1.0.0+computecanada
grpcio==1.51.3+computecanada
h5py==3.8.0+computecanada
idna==3.7
imageio==2.34.1+computecanada
imgaug==0.2.6
importlib-metadata==7.1.0
kornia==0.6.7
lightning-utilities==0.11.2+computecanada
loralib==0.1.1
Markdown==3.6
MarkupSafe==2.1.2+computecanada
multidict==6.0.5
networkx==2.8.7+computecanada
numpy==1.24.2+computecanada
oauthlib==3.2.2+computecanada
omegaconf==2.1.1
opencv-python-headless==4.6.0.66+computecanada
packaging==24.0
pandas==1.5.3+computecanada
Pillow==9.5.0+computecanada
Pillow-SIMD==9.0.0.post1+computecanada
protobuf==5.27.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydeprecate==0.3.1+computecanada
python-dateutil==2.9.0.post0+computecanada
pytorch-lightning==1.4.2+computecanada
pytz==2024.1+computecanada
PyWavelets==1.3.0+computecanada
PyYAML==6.0+computecanada
regex==2022.10.31+computecanada
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9+computecanada
scikit-image==0.19.3+computecanada
scipy==1.10.1+computecanada
six==1.15.0+computecanada
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.12.3+computecanada
tensorboard-data-server==0.7.2+computecanada
test-tube==0.7.5
tifffile==2022.8.12+computecanada
torch==1.12.1+computecanada
torchmetrics==0.6.0+computecanada
torchvision==0.13.1+torch1121.computecanada
tqdm==4.66.4+computecanada
typing-extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13+computecanada
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.1
